moved tests into utils
